Applyng DDD
