+++ src/modules/users/services/UpdateUserAvatarService.ts
@@ -1,11 +1,11 @@
 import { getRepository } from 'typeorm';
 import path from 'path';
 import fs from 'fs';
-import User from '../models/User';
 
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
 
-import uploadConfig from '../config/upload';
+import uploadConfig from '@config/upload';
+import User from '../infra/typeorm/entities/User';
 
 interface Request {
   user_id: string;
