+++ src/shared/infra/http/routes/index.ts
@@ -1,0 +1,12 @@
+import { Router } from 'express';
+import appointmentsRoutes from '@modules/appointments/infra/http/routes/appointments.routes';
+import usersRoutes from '@modules/users/infra/http/routes/users.routes';
+import sessionsRoutes from '@modules/users/infra/http/routes/sessions.routes';
+
+const routes = Router();
+
+routes.use('/appointments', appointmentsRoutes);
+routes.use('/users', usersRoutes);
+routes.use('/sessions', sessionsRoutes);
+
+export default routes;
